+++ publisher/src/main/scala/org/podval/tools/publish/page/DirectoryPage.scala
@@ -90,6 +90,5 @@ final class DirectoryPage(site: Site, path: Path) extends FullMarkupPage(site, p
     .filterNot(_.isInstanceOf[FacsimilePage])
     .filterNot(_.isInstanceOf[StoreIndexPage])
     .filterNot(_.isInstanceOf[AllEntitiesPage])
-    .filterNot(_.isInstanceOf[ReportPage])
     .filter(_.path.path.init == path.path.init)  // TODO unify with the Page.parent path calculations
     .sortBy(_.title.toLowerCase)